+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use seconds for logged durations.
+- Format floating point number (including durations) to 3 digits.
+
 ## [0.6.0] - 2024-01-17
 
 ### Added